fix: 修复 Windows 上点击托盘打开剪贴板窗口时，窗口位置配置未生效的问题
